Resolve clashing top-level given names (openable, showable, parserAggregable)
Two modules each defined a top-level `given openable` (bitumen, embarcadero)
re-exported into the `soundness` package, producing a duplicate top-level
definition when bundled into a fat jar. `showable` and `parserAggregable`
clashed the same way (jacinta with ypsiloid and breviloquence).

Move jacinta's `showable` and `parserAggregable` into `object Json.Ast`, where
they resolve via companion scope with no import — this alone clears both those
clashes, since jacinta was the common module. The bitumen/embarcadero
`openable` givens are JVM-only and key on an abstract external path, so their
format marker's companion (in platform-neutral core) can't host them; rename
them to `tarPathOpenable` / `imagePathOpenable` instead.

Remaining orphan top-level givens are tracked in #1632.
