+++ lib/jacinta/src/core/soundness_jacinta_core.scala
@@ -35,8 +35,7 @@ package soundness
 export
   jacinta
   . { dynamicJsonAccess, DynamicJsonEnabler, j, jp, Json, Json2, jsonConversion, JsonError,
-      JsonPointer, JsonPointerError, JsonPrimitive, JsonReader, NumberMode,
-      parserAggregable, showable }
+      JsonPointer, JsonPointerError, JsonPrimitive, JsonReader, NumberMode }
 
 package formatting:
   export jacinta.formatting.{indentedJsonFormatting, compactJsonFormatting}
